refactor(Shared): Move all the shareable content in the Shared folder
I'm moving the ECS all the old include folder, which is now split into subfolders, the DLLoader, the NetworkPacket and the SceneManager
