+++ Engine/Shared/Network/Packet/CMakeLists.txt
@@ -1,12 +1,12 @@
 cmake_minimum_required(VERSION 3.5)
 
 set(LIB r-type_packet)
-set(LIB_INCLUDE_PATH "include")
+set(LIB_INCLUDE_PATH "Include")
 
 set(CMAKE_CXX_STANDARD 20)
 
 set(SRC_LIB
-        ./NetworkPacket.cpp
+        NetworkPacket.cpp
 )
 
 add_library(${LIB} STATIC ${SRC_LIB})
